+++ asm/lib.rs
@@ -72,7 +72,8 @@ shims! {
     fn __udf();
     fn __wfe();
     fn __wfi();
-    fn __syscall(nr: u32, arg: u32) -> u32;
+    fn __sh_syscall(nr: u32, arg: u32) -> u32;
+    fn __bootstrap(msp: u32, rv: u32);
 }
 
 // v7m *AND* v8m.main, but *NOT* v8m.base