+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- New assembly methods `asm::sh_syscall`, `asm::bootstrap`, and
+  `asm::bootload`.
+
 ## [v0.7.0] - 2020-11-09
 
 ### Added